Move `QuadletOptions` into module
